requirements.txt
openai
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
python-dotenv
streamlit
pydeck